+++ Firmware/Application/app/app_version.h
@@ -1,0 +1,6 @@
+#ifndef APP_VERSION_H
+#define APP_VERSION_H
+
+#define APPLICATION_VERSION "0.3.0"
+
+#endif /* APP_VERSION_H */