+++ Firmware/Bootloader/app/boot_version.h
@@ -1,6 +1,6 @@
 #ifndef BOOT_VERSION_H
 #define BOOT_VERSION_H
 
-#define BOOTLOADER_VERSION "0.2.0"
+#define BOOTLOADER_VERSION "0.3.0"
 
 #endif /* BOOT_VERSION_H */